Promote drafts/ActionList2 to main/ActionList
