+++ docs/content/drafts/ActionMenu2.mdx
@@ -306,7 +306,7 @@ render(
 
 | Name                                     | Type                  |       Default       | Description                                                                                   |
 | :--------------------------------------- | :-------------------- | :-----------------: | :-------------------------------------------------------------------------------------------- |
-| children\*                               | `React.ReactElement[] | React.ReactElement` | Required. Recommended: [`ActionList`](/drafts/ActionList2)                                    |
+| children\*                               | `React.ReactElement[] | React.ReactElement` | Required. Recommended: [`ActionList`](/ActionList)                                            |
 | [OverlayProps](/Overlay#component-props) | -                     |          -          | Optional. Props to be spread on the internal [`AnchoredOverlay`](/AnchoredOverlay) component. |
 
 ## Status
@@ -336,6 +336,6 @@ render(
 
 ## Related components
 
-- [ActionList](/drafts/ActionList2)
+- [ActionList](/drafts/ActionList)
 - [SelectPanel](/SelectPanel)
 - [Button](/drafts/Button2)
