+++ docs/content/drafts/DropdownMenu2.mdx
@@ -312,7 +312,7 @@ Use `DropdownMenu` to select an option from a small list. If you’re looking fo
     description={
       <>
         Recommended:{' '}
-        <Link href="/drafts/ActionList2">
+        <Link href="/drafts/ActionList">
           <InlineCode>ActionList</InlineCode>
         </Link>
       </>
@@ -359,6 +359,6 @@ Use `DropdownMenu` to select an option from a small list. If you’re looking fo
 
 ## Related components
 
-- [ActionList](/drafts/ActionList2)
+- [ActionList](/drafts/ActionList)
 - [ActionMenu](/ActionMenu2)
 - [SelectPanel](/SelectPanel)
