+++ src/__tests__/ActionMenu2.test.tsx
@@ -4,7 +4,7 @@ import {axe, toHaveNoViolations} from 'jest-axe'
 import React from 'react'
 import theme from '../theme'
 import {ActionMenu} from '../ActionMenu2'
-import {ActionList} from '../ActionList2'
+import {ActionList} from '../ActionList'
 import {behavesAsComponent, checkExports, checkStoriesForAxeViolations} from '../utils/testing'
 import {BaseStyles, ThemeProvider, SSRProvider} from '..'
 import {SingleSelection, MixedSelection} from '../../src/stories/ActionMenu2/examples.stories'
