+++ src/__tests__/DropdownMenu2.test.tsx
@@ -4,7 +4,7 @@ import {toHaveNoViolations} from 'jest-axe'
 import React from 'react'
 import theme from '../theme'
 import {DropdownMenu} from '../DropdownMenu2'
-import {ActionList} from '../ActionList2'
+import {ActionList} from '../ActionList'
 import {behavesAsComponent, checkExports, checkStoriesForAxeViolations} from '../utils/testing'
 import {BaseStyles, ThemeProvider, SSRProvider} from '..'
 import '@testing-library/jest-dom'
