+++ src/drafts/index.ts
@@ -6,7 +6,6 @@
  */
 
 // Components
-export * from '../ActionList2'
 export * from '../Button2'
 export * from '../ActionMenu2'
 export * from '../DropdownMenu2'
