+++ src/index.ts
@@ -29,7 +29,15 @@ export {useConfirm} from './Dialog/ConfirmationDialog'
 // Components
 export {default as Radio} from './Radio'
 export type {RadioProps} from './Radio'
-
+export {ActionList} from './ActionList'
+export type {
+  ActionListProps,
+  ActionListGroupProps,
+  ActionListItemProps,
+  ActionListDescriptionProps,
+  ActionListLeadingVisualProps,
+  ActionListTrailingVisualProps
+} from './ActionList'
 export {ActionMenu} from './ActionMenu'
 export type {ActionMenuProps} from './ActionMenu'
 export {default as Autocomplete} from './Autocomplete'
