+++ src/stories/DropdownMenu2/fixtures.stories.tsx
@@ -3,7 +3,7 @@ import {Meta} from '@storybook/react'
 import {ThemeProvider} from '../..'
 import BaseStyles from '../../BaseStyles'
 import {DropdownMenu} from '../../DropdownMenu2'
-import {ActionList} from '../../ActionList2'
+import {ActionList} from '../../ActionList'
 import {Button} from '../../Button2'
 import Box from '../../Box'
 import Text from '../../Text'
