+++ src/stories/deprecated/ActionList.stories.tsx
@@ -14,11 +14,11 @@ import {
 import {Meta} from '@storybook/react'
 import React, {forwardRef} from 'react'
 import styled from 'styled-components'
-import {Label, ThemeProvider} from '..'
-import {ActionList as _ActionList} from '../deprecated/ActionList'
-import {Header} from '../deprecated/ActionList/Header'
-import BaseStyles from '../BaseStyles'
-import sx from '../sx'
+import {Label, ThemeProvider} from '../..'
+import {ActionList as _ActionList} from '../../deprecated/ActionList'
+import {Header} from '../../deprecated/ActionList/Header'
+import BaseStyles from '../../BaseStyles'
+import sx from '../../sx'
 
 const ActionList = Object.assign(_ActionList, {
   Header
